RulesTable clean up
